addition and subtraction implemented
multiplication stuff done with duals

arithmetic operations done

implemented unary negation

more dual num progress

backup

minor cleanups

backup

more fixes

backup

backup with more progress

more progress

most stuff with Dual done

demonstrating Scalar trait on Dual

initial prototype working

backu

adding some docs

new todos, some docs fixes

backup

more todos

more progress

almost there on the vector side, Vectorf64 should also be a trait

some more progress, still stuck

close to working vresion

basic version working after specializing to DVector

backup with slightly more progress

backup

backup for now

dual vec implemented

backup with some gradient progress

even more gradient progress

autodiff gradient working

gradient passing with two different types

gradient tests all passing

added a TODO

more TODOs, versioning updates

added some docs to dual_vector

more testing

backup

gradient actually working with multiple types

got vector valued deriv working with multiple types (hopefully)

more docs for get_sderivative

starting to fill out scalar-valued derivative tests

updated order of args

backup

still stuck

poly tests done

square root test done

tested product and quotient rules

hit another stopping point

got this thing working

exponential tests added

backup

power improvement

up to sin

backup for tonight

implementation of trig trait for Dual num

sin test

more trig tests

csc tests done

sec tested

finished testing trig functions

basic inverse trig tests

done with inverse trig tests

renamed test funcs

hyperbolic functions done

half of inverse hyperbolic functions done

finished inverse hyperbolic tests

renamed something

more progres

more docs updates

more progress

some more progress on docs

got rid of a TODO

+++ src/test_utils.rs
@@ -17,8 +17,8 @@ pub(crate) fn sqrt_deriv(x: f64) -> f64 {
 }
 
 #[cfg(test)]
-pub(crate) fn power_deriv(b: i32, x: f64) -> f64 {
-    (b as f64).powf(x) * (b as f64).ln()
+pub(crate) fn power_deriv(b: f64, x: f64) -> f64 {
+    b.powf(x) * b.ln()
 }
 
 #[cfg(test)]
